Added a UI for the editor page
flutter_code_editor package has been modified by me to support e.g. custom line numbers (with avatar icon), to fix a problem with the scrollbar, also the problem with pressing the enter key and with the position for a popup widget. This package is under the Apache 2.0 license, so there should be no problem with me doing this and making a fork.